Rename the fields of the data entered in the section autoreport of the website
Now we enter the hostname (not a description string anymore), and a hash string (not an ip anymore)
The plugin will detect the ip of the remote client claiming to report a match.
Then, it will get the hostname associated with this ip and it will compare it to those that are
recorded in the database. Then it will check the hash string sent by the remote reporter.
hostname and hash string must match in order to allow any reporting.

This commit is only the website side of this development.